utils/requirements.txt
jinja2>=3.0.0
tree-sitter>=0.25.0
tree-sitter-rust>=0.24.0
PyYAML>=6.0